inference/requirements.txt
prometheus_client==0.22.1
fastapi==0.116.1
uvicorn==0.35.0
redis==5.0.4
numpy==1.26.4
pandas==2.1.4
pyarrow==18.1.0
mlflow==3.2.0
boto3
requests
scikit-learn
joblib
pydantic
protobuf==6.32.0
python-json-logger
prometheus-fastapi-instrumentator
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi